+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Support for parsing a variable in the index.html file during cloning (e.g. SEO metadata).
+
 ## [0.1.0] - 2023-01-27
 
 ### Added